+++ pyupdater/__init__.py
@@ -57,9 +57,9 @@ except ImportError:  # pragma: no cover
     pyi_version = (0, 0, 0)
 
 if pyi_version < (2, 1, 0):  # pragma: no cover
-    raise pyupdater.exceptions.PyiUpdaterError(u'Must have at least '
-                                               u'PyInstaller v2.1',
-                                               expected=True)
+    raise pyupdater.exceptions.PyUpdaterError(u'Must have at least '
+                                              u'PyInstaller v2.1',
+                                              expected=True)
 
 
 PyUpdater = pyupdater.core.Core